// src/app/models/storyRequestDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StoryRequestDto { 
    title?: string;
    description?: string;
    mainCharacters?: Array<string>;
    category?: StoryRequestDto.CategoryEnum;
    tags?: Array<string>;
    targetAudience?: StoryRequestDto.TargetAudienceEnum;
    language?: StoryRequestDto.LanguageEnum;
    copyright?: StoryRequestDto.CopyrightEnum;
    mature?: boolean;
    coverImageUrl?: string;
}
export namespace StoryRequestDto {
    export const CategoryEnum = {
        Action: 'ACTION',
        Adventure: 'ADVENTURE',
        Chicklit: 'CHICKLIT',
        Fanfiction: 'FANFICTION',
        Fantasy: 'FANTASY',
        GeneralFiction: 'GENERAL_FICTION',
        HistoricalFiction: 'HISTORICAL_FICTION',
        Horror: 'HORROR',
        Humor: 'HUMOR',
        MysteryThriller: 'MYSTERY_THRILLER',
        NonFiction: 'NON_FICTION',
        Paranormal: 'PARANORMAL',
        Poetry: 'POETRY',
        Random: 'RANDOM',
        Romance: 'ROMANCE',
        ScienceFiction: 'SCIENCE_FICTION',
        ShortStory: 'SHORT_STORY',
        Spiritual: 'SPIRITUAL',
        TeenFiction: 'TEEN_FICTION',
        Vampire: 'VAMPIRE',
        Werewolf: 'WEREWOLF'
    } as const;
    export type CategoryEnum = typeof CategoryEnum[keyof typeof CategoryEnum];
    export const TargetAudienceEnum = {
        YoungAdult: 'YOUNG_ADULT',
        NewAdult: 'NEW_ADULT',
        Adult: 'ADULT'
    } as const;
    export type TargetAudienceEnum = typeof TargetAudienceEnum[keyof typeof TargetAudienceEnum];
    export const LanguageEnum = {
        Polski: 'POLSKI',
        Arabic: 'ARABIC',
        Czech: 'CZECH',
        Deutsch: 'DEUTSCH',
        Greek: 'GREEK',
        English: 'ENGLISH',
        Spanish: 'SPANISH',
        Farsi: 'FARSI',
        Filipino: 'FILIPINO',
        French: 'FRENCH',
        Hindi: 'HINDI',
        BahasaIndonesia: 'BAHASA_INDONESIA',
        Italiano: 'ITALIANO',
        Hebrew: 'HEBREW',
        Magyar: 'MAGYAR',
        BahasaMelayu: 'BAHASA_MELAYU',
        Nederlands: 'NEDERLANDS',
        Portuguese: 'PORTUGUESE',
        Romanian: 'ROMANIAN',
        Russian: 'RUSSIAN',
        Slovak: 'SLOVAK',
        Serbian: 'SERBIAN',
        Turkish: 'TURKISH',
        Ukrainian: 'UKRAINIAN',
        Vietnamese: 'VIETNAMESE',
        Other: 'OTHER'
    } as const;
    export type LanguageEnum = typeof LanguageEnum[keyof typeof LanguageEnum];
    export const CopyrightEnum = {
        AllRightsReserved: 'ALL_RIGHTS_RESERVED',
        PublicDomain: 'PUBLIC_DOMAIN',
        CreativeCommonsAttribution: 'CREATIVE_COMMONS_ATTRIBUTION',
        CreativeCommonsAttributionNonCommercial: 'CREATIVE_COMMONS_ATTRIBUTION_NON_COMMERCIAL',
        CreativeCommonsAttributionNonCommercialNoDerivs: 'CREATIVE_COMMONS_ATTRIBUTION_NON_COMMERCIAL_NO_DERIVS',
        CreativeCommonsAttributionNonCommercialShareAlike: 'CREATIVE_COMMONS_ATTRIBUTION_NON_COMMERCIAL_SHARE_ALIKE',
        CreativeCommonsAttributionShareAlike: 'CREATIVE_COMMONS_ATTRIBUTION_SHARE_ALIKE',
        CreativeCommonsAttributionNoDerivs: 'CREATIVE_COMMONS_ATTRIBUTION_NO_DERIVS'
    } as const;
    export type CopyrightEnum = typeof CopyrightEnum[keyof typeof CopyrightEnum];
}